Fix first character indent
Fixes #54

Sets padding-left to 0 for code element to override themes, scoped to
just the code block. Can be tested in TwentySixteen, but likely others.

This change also cleans up the CSS that was being included but not used.
Some change in the past changed the class reference and the definitions
in the file were not in any scope for the block.