Add default color for wp-block-code code
The TwentyTwentyOne theme specifies a dark color for code blocks when
dark mode support is turned on, but in a light mode. This overrides
what the theme is for the syntax block.

For example, the default code block uses a black background, so if any
text is entered that is not a keyword, string, etc.. it gets the
default color from the theme.

This change adds a default color for pre.wp-block-code code to the
themes shipped with the plugin.

Fixes #98