CSS clean up, always apply background to block
Fixes background color if no language specified.

The background color was specified in the CSS to avoid applying to
inline code blocks, this is from the default Prism CSS. This had the
side effect of not applying the background color unless a language was
specified.

The rub is that using a Code block will always wrap the code in a pre
and code tags, so using the code-syntax-block you can not insert an
inline code block.

So the CSS was updated to only work against pre.wp-block-code which will
apply to all code blocks regardless of language specified or not.

Fixes #106, #94