Refactor for only useBlockProps support (WP 5.6+)
Previous WordPress versions used different markup for the code block.
This simplifies the plugin to support last three versions of WP, so
the update will require WP 5.6 or newer.

It is easier to update the plugin with new feature without needing
to backport to older versions.

This PR also includes a minor feature of adding a indicator label in
editor when line numbers are set to show.